# building-manager/technical/openapi.yaml
openapi: 3.0.3
info:
  title: Building Manager API
  description: API for the Building Manager (Upravnik Zgrade) application
  version: 1.0.0
  contact:
    name: Building Manager Team
    email: support@upravnik.app
servers:
  - url: https://api.upravnik.app/api/v1
    description: Production
  - url: https://staging-api.upravnik.app/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Development
...
paths:
...
  /buildings:
    get:
      tags: [Buildings]
      summary: List user's buildings
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        '200':
          description: List of buildings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingListResponse'
    post:
      tags: [Buildings]
      summary: Create building
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildingRequest'
      responses:
        '201':
          description: Building created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '400':
          $ref: '#/components/responses/ValidationError'
  /buildings/{buildingId}:
    get:
      tags: [Buildings]
      summary: Get building details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
      responses:
        '200':
          description: Building details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Buildings]
      summary: Update building
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBuildingRequest'
      responses:
        '200':
          description: Building updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Buildings]
      summary: Archive building
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
      responses:
        '200':
          description: Building archived
        '404':
          $ref: '#/components/responses/NotFound'
  /buildings/{buildingId}/units:
    get:
      tags: [Units]
      summary: List building units
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: floor
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UnitStatus'
      responses:
        '200':
          description: List of units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitListResponse'
    post:
      tags: [Units]
      summary: Create unit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnitRequest'
      responses:
        '201':
          description: Unit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unit'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /buildings/{buildingId}/residents:
    get:
      tags: [Residents]
      summary: List residents
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          description: Search by name, unit number, or phone
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [owner, tenant, family]
        - name: unitId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of residents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentListResponse'
    post:
      tags: [Residents]
      summary: Create resident
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResidentRequest'
      responses:
        '201':
          description: Resident created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /buildings/{buildingId}/keys:
    get:
      tags: [Keys]
      summary: List keys
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/KeyType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/KeyStatus'
        - name: unitId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyListResponse'
    post:
      tags: [Keys]
      summary: Create key
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeyRequest'
      responses:
        '201':
          description: Key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Key'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /buildings/{buildingId}/announcements:
    get:
      tags: [Announcements]
      summary: List announcements
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AnnouncementType'
      responses:
        '200':
          description: List of announcements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementListResponse'
    post:
      tags: [Announcements]
      summary: Create announcement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/buildingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnnouncementRequest'
      responses:
        '201':
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    buildingId:
      name: buildingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
...
    Building:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        streetAddress:
          type: string
        streetNumber:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        yearBuilt:
          type: integer
        totalFloors:
          type: integer
        hasElevator:
          type: boolean
        hasParking:
          type: boolean
        totalAreaM2:
          type: number
        isLegalEntity:
          type: boolean
        legalEntityName:
          type: string
        mbNumber:
          type: string
        pib:
          type: string
        unitCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    BuildingListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Building'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'
    CreateBuildingRequest:
      type: object
      required: [streetAddress, streetNumber]
      properties:
        name:
          type: string
        streetAddress:
          type: string
        streetNumber:
          type: string
        city:
          type: string
        postalCode:
          type: string
        yearBuilt:
          type: integer
        totalFloors:
          type: integer
        hasElevator:
          type: boolean
        hasParking:
          type: boolean
        totalAreaM2:
          type: number
    UpdateBuildingRequest:
      type: object
      properties:
        name:
          type: string
        isLegalEntity:
          type: boolean
        legalEntityName:
          type: string
        mbNumber:
          type: string
        pib:
          type: string
        bankAccount:
          type: string
        registrationDate:
          type: string
          format: date

    # -------------------------------------------------------------------------
    # Unit
    # -------------------------------------------------------------------------
    Unit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        number:
          type: string
        floor:
          type: integer
        entrance:
          type: string
        areaM2:
          type: number
        ownershipShare:
          type: number
        votingShare:
          type: number
        type:
          $ref: '#/components/schemas/UnitType'
        status:
          $ref: '#/components/schemas/UnitStatus'
        ownerCount:
          type: integer
        residentCount:
          type: integer
...
    UnitListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Unit'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'
    UnitType:
      type: string
      enum: [APARTMENT, STUDIO, LOFT, COMMERCIAL, OFFICE, GARAGE, STORAGE, MIXED_USE]
    UnitStatus:
      type: string
      enum: [OCCUPIED_OWNER, OCCUPIED_FAMILY, OCCUPIED_TENANT, VACANT, UNDER_RENOVATION]
    CreateUnitRequest:
      type: object
      required: [number, floor, areaM2, ownershipShare]
      properties:
        number:
          type: string
        floor:
          type: integer
        entrance:
          type: string
        position:
          type: string
        areaM2:
          type: number
        roomCount:
          type: integer
        bedroomCount:
          type: integer
        bathroomCount:
          type: integer
        ownershipShare:
          type: number
        type:
          $ref: '#/components/schemas/UnitType'
...
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        preferredName:
          type: string
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        emails:
          type: array
          items:
            $ref: '#/components/schemas/Email'
        emergencyName:
          type: string
        emergencyPhone:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
...
    Phone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        type:
          $ref: '#/components/schemas/PhoneType'
        isPrimary:
          type: boolean
        isVerified:
          type: boolean
    PhoneType:
      type: string
      enum: [MOBILE, HOME, WORK, OTHER]
    Email:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
          format: email
        type:
          type: string
        isPrimary:
          type: boolean
        isVerified:
          type: boolean
...
    ResidentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Person'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'
    CreateResidentRequest:
      type: object
      required: [firstName, lastName, unitId, role]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        unitId:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, tenant, family]
        phone:
          type: string
        email:
          type: string
          format: email
        emergencyName:
          type: string
        emergencyPhone:
          type: string
        ownershipShare:
          type: number
          description: Required if role is owner
        moveInDate:
          type: string
          format: date
        leaseStart:
          type: string
          format: date
          description: Required if role is tenant
        leaseEnd:
          type: string
          format: date
...
    Key:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        unitId:
          type: string
          format: uuid
        identifier:
          type: string
        type:
          $ref: '#/components/schemas/KeyType'
        description:
          type: string
        status:
          $ref: '#/components/schemas/KeyStatus'
        currentTransaction:
          $ref: '#/components/schemas/KeyTransaction'
...
    KeyType:
      type: string
      enum: [ENTRANCE_MAIN, ENTRANCE_SIDE, UNIT, STORAGE, GARAGE, COMMON_ROOM, MAILBOX, OTHER]
    KeyStatus:
      type: string
      enum: [AVAILABLE, CHECKED_OUT, LOST, DAMAGED]
    KeyListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Key'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'
    CreateKeyRequest:
      type: object
      required: [type, identifier]
      properties:
        unitId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/KeyType'
        identifier:
          type: string
        description:
          type: string
        lockManufacturer:
          type: string
        lockModel:
          type: string
...
    Announcement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        type:
          $ref: '#/components/schemas/AnnouncementType'
        targetType:
          $ref: '#/components/schemas/TargetType'
        channels:
          type: array
          items:
            $ref: '#/components/schemas/NotificationChannel'
        scheduledFor:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
        createdById:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    AnnouncementType:
      type: string
      enum: [GENERAL, EMERGENCY, MAINTENANCE, MEETING, PAYMENT_REMINDER]
    TargetType:
      type: string
      enum: [ALL, OWNERS_ONLY, TENANTS_ONLY, SPECIFIC_UNITS]
    NotificationChannel:
      type: string
      enum: [PUSH, SMS, EMAIL]
    AnnouncementListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Announcement'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'
    CreateAnnouncementRequest:
      type: object
      required: [title, content, type, targetType, channels]
      properties:
        title:
          type: string
          maxLength: 100
        content:
          type: string
          maxLength: 1000
        type:
          $ref: '#/components/schemas/AnnouncementType'
        targetType:
          $ref: '#/components/schemas/TargetType'
        targetUnitIds:
          type: array
          items:
            type: string
            format: uuid
        channels:
          type: array
          items:
            $ref: '#/components/schemas/NotificationChannel'
        scheduledFor:
          type: string
          format: date-time
...